+++ earthdaily/earthdatastore/__init__.py
@@ -13,12 +13,11 @@ from pystac.item_collection import ItemCollection
 from pystac_client import Client
 
 from . import _scales_collections, cube_utils, mask
-from .cube_utils import datacube, metacube
+from .cube_utils import datacube, metacube, _datacubes
 
 __all__ = ["datacube", "metacube", "xr"]
 
 logging.getLogger("earthdaily-earthdatastore")
-
 
 def post_query_items(items, query):
     items_ = []
@@ -397,7 +396,8 @@ class Auth:
         query.update(ag_query)
         search_kwargs[target_param] = query
         return search_kwargs
-
+    
+    @_datacubes
     def datacube(
         self,
         collections: str | list,
@@ -450,7 +450,7 @@ class Auth:
         )
 
         xcal_items = None
-        if isinstance(cross_calibration_collection, str):
+        if isinstance(cross_calibration_collection, str) and cross_calibration_collection != collections[0]:
             try:
                 xcal_items = self.search(
                     collections="eda-cross-calibration",