+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.1]
 
+### Added
+
+- Auth datacube function now manage multiple collections
 
 ## [0.0.1-rc9] - 2023-12-12